fix(server): tag uint32 wire fields format:int64 to stop SDK narrowing (#90)
Huma's reflection maps Go uint32 → OpenAPI integer/int32 by default,
but Meshtastic node nums and packet ids are chip-derived uint32 values
that routinely exceed 2^31-1 — so the generated SDK fails to decode
real-radio payloads with "cannot unmarshal number into int32".

Annotate the wire-visible uint32 fields (MyNodeNum, BatteryLevel,
PacketID, ReplyID across RadioSummary / SessionSnapshot /
SendMessageRequest / SendMessageResult / PingResult / TracerouteResult)
with format:"int64" minimum:"0" so the spec emits int64 and oapi-codegen
generates Go int64 instead of int32. Internal command shapes
(model/commands.go) are unaffected — they never hit the HTTP wire.

Regenerated internal/sdk/gen/{api.yaml,client.gen.go} reflects the new
shapes; updated sdk/remote.go to widen the in-flight int32 conversion
for ReplyId on the SendMessage POST.

+++ internal/sdk/remote.go
@@ -187,7 +187,7 @@ func (r *Remote) Send(cmd mdl.Command) (uint32, bool) {
 			Text:    c.Text,
 		}
 		if c.ReplyID != 0 {
-			rid := int32(c.ReplyID)
+			rid := int64(c.ReplyID)
 			body.ReplyId = &rid
 		}
 		// Bound the outbound POST so an unreachable daemon doesn't
